ovn-controller-vtep: Add binding module.
This commit adds the binding module to ovn-controller-vtep.  The
module will scan through the Binding table in ovnsb.  If there is
a binding for the logical port on vtep gw, sets the binding's chassis
column to the corresponding vtep gw's chassis.

Note, each logical datapath can only have up to one logical port from
each 'vlan_map' (in Gateway table) attached to it.

Signed-off-by: Alex Wang <alexw@nicira.com>