ovn-controller-vtep: Add pipeline module.
This commit adds the pipeline module to ovn-controller-vtep.  The
module will scan through the Binding table, and create the
Ucast_Macs_Remote in vtep database for each logical port's MACs in
the same logical datapath (same as logical switch for vtep).  Also
the module will create the physical locators to each HV chassis in
the same logical datapath.

Signed-off-by: Alex Wang <alexw@nicira.com>