ovn-sb: Remove the "Gateway" table from the ovn-sb schema.
In a gateway like the VTEP L2 gateway, physical vlans belonging to
the same logical network form a "logical switch".  Each logical switch
has a dedicated tunnel key and will keep records of all MACs learned
from the owned vlans.  So user can just send packet to a "logical
switch" and the gateway will figure out the output port and vlan tag
automatically.

Therefore, it is not really necessary to keep record of the vlan map
for each gateway physical port in the OVN_Southbound database using
"gateway_ports".

Thusly, this commit removes the "Gateway" table from the OVN_Southbound
database.  In the "Chassis" table, the "gateway_ports" column is replaced
by "logical_switches" column which maps the logical switch name in the
gateway to a logical port name that exists in the OVN_Northbound database's
"Logical_Port" table.

Signed-off-by: Alex Wang <alexw@nicira.com>

---
->V3:
- Realize that the Gateway table is not needed.